requirements.txt
boto3>=1.34.0
openai>=1.40.0
pyyaml>=6.0
pandas>=2.1.0
tqdm>=4.66.0

